chore: move cert itests to spec tests (#23607)
