+++ tests/specs/cert/cafile_bundle/__test__.jsonc
@@ -1,0 +1,11 @@
+{
+  "tempDir": true,
+  "steps": [{
+    "args": "bundle --cert RootCA.pem https://localhost:5545/subdir/mod1.ts mod1.bundle.js",
+    "flaky": true,
+    "output": "[WILDCARD]"
+  }, {
+    "args": "run --quiet --check test.js",
+    "output": "[WILDCARD]Hello\n"
+  }]
+}
