+++ tests/specs/cert/cafile_bundle/test.js
@@ -1,0 +1,2 @@
+import { printHello3 } from "./mod1.bundle.js";
+printHello3();
