+++ tests/specs/cert/cafile_ts_fetch/cafile_ts_fetch.ts.out
@@ -1,0 +1,2 @@
+[WILDCARD]
+Hello
