+++ tests/specs/cert/cafile_ts_fetch_unsafe_ssl/cafile_ts_fetch.ts
@@ -1,0 +1,3 @@
+fetch("https://localhost:5545/cert/cafile_ts_fetch.ts.out")
+  .then((r) => r.text())
+  .then((t) => console.log(t.trimEnd()));
