+++ tests/specs/cert/cafile_url_imports/__test__.jsonc
@@ -1,0 +1,5 @@
+{
+  "args": "run --quiet --reload --cert RootCA.pem cafile_url_imports.ts",
+  "flaky": true,
+  "output": "cafile_url_imports.ts.out"
+}
