+++ tests/specs/cert/cafile_url_imports/cafile_url_imports.ts
@@ -1,0 +1,3 @@
+import { printHello } from "https://localhost:5545/subdir/mod2.ts";
+printHello();
+console.log("success");
