+++ tests/specs/cert/deno_land_unsafe_ssl/__test__.jsonc
@@ -1,0 +1,5 @@
+{
+  "flaky": true,
+  "args": "run --quiet --reload --allow-net --unsafely-ignore-certificate-errors=deno.land deno_land_unsafe_ssl.ts",
+  "output": "deno_land_unsafe_ssl.ts.out"
+}
