+++ tests/specs/cert/deno_land_unsafe_ssl/deno_land_unsafe_ssl.ts
@@ -1,0 +1,2 @@
+const r = await fetch("https://github.com");
+console.log(r.status);
