+++ tests/specs/cert/ip_address_unsafe_ssl/__test__.jsonc
@@ -1,0 +1,4 @@
+{
+  "args": "run --quiet --reload --allow-net --unsafely-ignore-certificate-errors=1.1.1.1 ip_address_unsafe_ssl.ts",
+  "output": "ip_address_unsafe_ssl.ts.out"
+}
