--- tests/testdata/cert/deno_land_unsafe_ssl.ts
@@ -1,2 +1,0 @@
-const r = await fetch("https://google.com");
-console.log(r.status);
