+++ tests/util/server/src/macros.rs
@@ -66,30 +66,6 @@ macro_rules! itest(
 );
 
 #[macro_export]
-macro_rules! itest_flaky(
-($name:ident {$( $key:ident: $value:expr,)*})  => {
-  #[flaky_test::flaky_test]
-  fn $name() {
-    $crate::timeout!();
-    let test = $crate::CheckOutputIntegrationTest {
-      $(
-        $key: $value,
-       )*
-      .. Default::default()
-    };
-    let output = test.output();
-    output.assert_exit_code(test.exit_code);
-    if !test.output.is_empty() {
-      assert!(test.output_str.is_none());
-      output.assert_matches_file(test.output);
-    } else {
-      output.assert_matches_text(test.output_str.unwrap_or(""));
-    }
-  }
-}
-);
-
-#[macro_export]
 macro_rules! context(
 ({$( $key:ident: $value:expr,)*})  => {
   $crate::TestContext::create($crate::TestContextOptions {
@@ -98,21 +74,6 @@ macro_rules! context(
       )*
     .. Default::default()
   })
-}
-);
-
-#[macro_export]
-macro_rules! itest_steps(
-($name:ident {$( $key:ident: $value:expr,)*})  => {
-  #[test]
-  fn $name() {
-    ($crate::CheckOutputIntegrationTestSteps {
-      $(
-        $key: $value,
-       )*
-      .. Default::default()
-    }).run()
-  }
 }
 );
